requirements.txt
fastapi
uvicorn[standard]
jsonpickle~=1.4.2
pyyaml~=5.3.1
pymongo~=3.12.0
passlib~=1.7.4
httpx~=0.16.1
dnspython
argon2-cffi
python-jose[cryptography]
starlette~=0.13.6
jinja2>=2.11.2
jwt